refactor code to re-order functions in their related package and files
