+++ pkg/steps/template_test.go
@@ -368,74 +368,3 @@ func TestInjectLabelsToTemplate(t *testing.T) {
 		})
 	}
 }
-
-func TestPodHasStarted(t *testing.T) {
-	for _, tc := range []struct {
-		name     string
-		pod      coreapi.Pod
-		expected bool
-	}{{
-		name: "pod is pending",
-		pod: coreapi.Pod{
-			Status: coreapi.PodStatus{Phase: coreapi.PodPending},
-		},
-	}, {
-		name: "pod is pending, init containers are not running",
-		pod: coreapi.Pod{
-			Status: coreapi.PodStatus{
-				Phase: coreapi.PodPending,
-				InitContainerStatuses: []coreapi.ContainerStatus{{
-					Name: "waiting",
-					State: coreapi.ContainerState{
-						Waiting: &coreapi.ContainerStateWaiting{},
-					},
-				}, {
-					Name: "terminated",
-					State: coreapi.ContainerState{
-						Terminated: &coreapi.ContainerStateTerminated{},
-					},
-				}},
-			},
-		},
-	}, {
-		name: "pod is pending, infrastructure init container is running",
-		pod: coreapi.Pod{
-			Status: coreapi.PodStatus{
-				Phase: coreapi.PodPending,
-				InitContainerStatuses: []coreapi.ContainerStatus{{
-					Name: "cp-secret-wrapper",
-					State: coreapi.ContainerState{
-						Running: &coreapi.ContainerStateRunning{},
-					},
-				}},
-			},
-		},
-	}, {
-		name: "pod is pending, init container is running",
-		pod: coreapi.Pod{
-			Status: coreapi.PodStatus{
-				Phase: coreapi.PodPending,
-				InitContainerStatuses: []coreapi.ContainerStatus{{
-					Name: "running",
-					State: coreapi.ContainerState{
-						Running: &coreapi.ContainerStateRunning{},
-					},
-				}},
-			},
-		},
-		expected: true,
-	}, {
-		name: "pod is running",
-		pod: coreapi.Pod{
-			Status: coreapi.PodStatus{Phase: coreapi.PodRunning},
-		},
-		expected: true,
-	}} {
-		t.Run(tc.name, func(t *testing.T) {
-			ret := podHasStarted(&tc.pod)
-			if ret != tc.expected {
-				t.Fatalf("got %v, want %v", ret, tc.expected)
-			}
-		})
-	}
-}
